RENO-2221: Update url patterns (#140)
* first draft of updating url patterns

* added online resources specific page container and refactored existing pgs
