+++ next.config.js
@@ -15,15 +15,16 @@ module.exports = (nextConfig = {}) => {
       return [];
     },
     // @TODO Not sure we need this.
-    async redirects() {
+    /*async redirects() {
       return [
         {
           source: '/collections/articles-databases/:slug',
-          destination: '/research/online-resources/:slug',
+          destination: '/research/collections/online-resources/:slug',
           permanent: true,
         },
       ]
     },
+    */
     webpack(config, options) {
       const { dir } = options
